Update var_EDU.R
The issues were related on how R deals with Missing

(1) When use case_when in R, the conditions are ordered, so it does not replace values for observations that matched any condition before.

So if you  have obs  i with var1=1 & var2= NA

case_when ( var1<1 ~ 0,

                     is.na(var2) ~NA)

it won't  assign NA to obs i.

To adjust that, I moved all is.na(cond) to the beginning of case_when

(2) On using ifelse is the opposite, it will assign 0 to all obs that does not match the condition, including the NA. So I added an additional line in all ifelse to be sure that NA values are replaced.

(3) adjusted age_group to age_18_23 instead of age_18_24 as cecilia asked in the email.